requirements.txt
dify_plugin==0.3.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-core==2.25.0
googleapis-common-protos==1.70.0
